Move TCServerInfo into its own files
